+++ http/src/main/java/io/metaloom/qdrant/client/http/model/cluster/operation/MoveShardOperation.java
@@ -1,29 +1,31 @@
 package io.metaloom.qdrant.client.http.model.cluster.operation;
 
+import java.math.BigInteger;
+
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class MoveShardOperation extends AbstractShardOperation {
 
 	@JsonProperty("to_peer_id")
-	private long toPeerId;
+	private BigInteger toPeerId;
 
 	@JsonProperty("from_peer_id")
-	private long fromPeerId;
+	private BigInteger fromPeerId;
 
-	public long getToPeerId() {
+	public BigInteger getToPeerId() {
 		return toPeerId;
 	}
-	
-	public MoveShardOperation setToPeerId(long toPeerId) {
+
+	public MoveShardOperation setToPeerId(BigInteger toPeerId) {
 		this.toPeerId = toPeerId;
 		return this;
 	}
 
-	public long getFromPeerId() {
+	public BigInteger getFromPeerId() {
 		return fromPeerId;
 	}
-	
-	public MoveShardOperation setFromPeerId(long fromPeerId) {
+
+	public MoveShardOperation setFromPeerId(BigInteger fromPeerId) {
 		this.fromPeerId = fromPeerId;
 		return this;
 	}
